backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlglot==25.24.5
python-multipart==0.0.9
google-generativeai==0.8.1
